Remove wakeup symbols from libupower-glib3.symbols
Strictly speaking, the removal of the wakeup API constitutes an API/ABI
break and the soname of the library should have been bumped. As upower
appears to be the only user of that API though, it should be relatively
safe without an explicit soname bump.

To ensure that both upower and gir1.2-upowerglib-1.0 are upgraded along
with libupower-glib3, add versioned Breaks against those packages to
libupower-glib3.